+++ src/full/Agda/Syntax/Translation/ConcreteToAbstract.hs
@@ -60,6 +60,7 @@ import Agda.Syntax.Concrete.Fixity (DoWarn(..))
 import Agda.Syntax.Notation
 import Agda.Syntax.Scope.Base as A
 import Agda.Syntax.Scope.Monad
+import Agda.Syntax.Scope.UnusedImports (lookedupName)
 import Agda.Syntax.Translation.AbstractToConcrete (ToConcrete, ConOfAbs)
 import Agda.Syntax.DoNotation
 import Agda.Syntax.IdiomBrackets
@@ -564,6 +565,7 @@ instance ToAbstract MaybeOldQName where
   type AbsOfCon MaybeOldQName = Maybe A.Expr
   toAbstract (MaybeOldQName (OldQName x ns)) = do
     qx <- resolveName' allKindsOfNames ns x
+    lookedupName x qx -- Mark this name as used for the unused-imports analysis
     reportSLn "scope.name" 30 $ "resolved " ++ prettyShow x ++ ": " ++ prettyShow qx
     case qx of
       VarName x' _         -> return $ Just $ A.Var x'
@@ -1579,7 +1581,10 @@ importPrimitives = do
           ]
         directives          = ImportDirective noRange (Using usingDirective) [] [] Nothing
         importAgdaPrimitive = [C.Import (C.DoOpen empty) empty agdaPrimitiveName empty directives]
-    toAbstract (Declarations importAgdaPrimitive)
+    -- We don't want UnusedImports warnings when importing the primitives,
+    -- since the open statement was not assembled by the user.
+    locallyTCState stWarningSet (Set.\\ unusedImportsWarnings) do
+      toAbstract (Declarations importAgdaPrimitive)
 
 -- | runs Syntax.Concrete.Definitions.niceDeclarations on main module
 niceDecls :: DoWarn -> [C.Declaration] -> ([NiceDeclaration] -> ScopeM a) -> ScopeM a