+++ src/full/Agda/TypeChecking/Pretty/Warning.hs
@@ -320,6 +320,13 @@ prettyWarning = \case
       pwords "Invalid character literal" ++ [text $ show c] ++
       pwords "(surrogate code points are not supported)"
 
+    UnusedImports m Nothing -> fsep $
+      pwords "Redundant opening of" ++ [ prettyTCM m ]
+
+    UnusedImports m (Just xs) -> fsep $
+      [ "Opening", prettyTCM m ] ++ pwords "brings the following unused names into scope:"
+      ++ fmap prettyTCM (List1.toList xs)
+
     UselessPragma _r d -> return d
 
     SafeFlagPostulate q -> fsep $