+++ src/full/Agda/Utils/Set.hs
@@ -1,7 +1,7 @@
 module Agda.Utils.Set (module Agda.Utils.Set, module Data.Set) where
 
 import Data.Set
-import Data.Set.Internal
+import Data.Set.Internal (Set(Tip, Bin))
 
 -- Adapted from rejected pull request to containers
 -- https://github.com/haskell/containers/pull/291