+++ test/test-suite-covers-warnings.sh
@@ -17,7 +17,11 @@ COVERED=$TMPDIR/covered.txt
 # Those are printed by `agda --help=warning` at line beginnings and follow
 # the camel-case convention, with at least two capital letters.
 #
-${AGDA_BIN:-agda} --help=warning | sed -nr 's/^([A-Z][a-z]+[A-Z][A-Za-z]+).*/\1/p' | sort > $HELPWARNS
+${AGDA_BIN:-agda} --help=warning | sed -nr 's/^([A-Z][a-z]+[A-Z][A-Za-z]+).*/\1/p' | sort | uniq > $HELPWARNS
+
+if [[ "$1" == "--debug" ]]; then
+  cat $HELPWARNS
+fi
 
 # Warnings we do not need to cover by the testsuite.
 #